Split apart large header
