+++ inst/include/dust/r/dust.hpp
@@ -20,6 +20,7 @@
 
 #include "dust/r/gpu.hpp"
 #include "dust/r/helpers.hpp"
+#include "dust/r/helpers_ode.hpp"
 #include "dust/r/random.hpp"
 
 namespace dust {
